Module QCheck.Observable

Observables are usable as arguments for random functions. The random function will observe its arguments in a way that is determined from the observable instance.

type -'a t

An observable for 'a, packing a printer and other things.

val equal : 'a t -> 'a -> 'a -> bool

val hash : 'a t -> 'a -> int

val print : 'a t -> 'a Print.t

val unit : unit t

val bool : bool t

val int : int t

val int32 : int32 t

val int64 : int64 t

val float : float t

val string : string t

val bytes : bytes t

val char : char t

val make : ?eq:('a -> 'a -> bool) -> ?hash:('a -> int) -> 'a Print.t -> 'a t

val map : ('a -> 'b) -> 'b t -> 'a t

val option : 'a t -> 'a option t

val result : 'a t -> 'e t -> ('a, 'e) Stdlib.result t

val list : 'a t -> 'a list t

val array : 'a t -> 'a array t

val pair : 'a t -> 'b t -> ('a * 'b) t

val triple : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t

val quad : 'a t -> 'b t -> 'c t -> 'd t -> ('a * 'b * 'c * 'd) t