Module QCheck2.Fn

A utility module of helpers for printing, shrinking, and applying generated function values.

val print : 'f fun_ Print.t

print f prints the implementation of generated function f.

The implementation always contains a default case, represented as _.

Note that printing a function before it was called in the test may not print the full implementation.

val apply : 'f fun_ -> 'f

apply f returns the underlying function to be used in tests. This is an alias for deconstructing as documented in fun_.