Module QCheck.Test

Module related to individual tests.

since

NEXT_RELEASE most of it moved to {!QCheck2}, and the type ['a cell] was made a private implementation detail.

val fail_report : string -> 'a

Fail the test with some additional message that will be reported.

since

0.7

val fail_reportf : ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'a

Format version of fail_report

since

0.7

val make_cell : ?⁠if_assumptions_fail:([ `Fatal | `Warning ] * float) -> ?⁠count:int -> ?⁠long_factor:int -> ?⁠max_gen:int -> ?⁠max_fail:int -> ?⁠small:('a -> int) -> ?⁠name:string -> 'a arbitrary -> ('a -> bool) -> 'a QCheck2.Test.cell

make_cell arb prop builds a test that checks property prop on instances of the generator arb.

parameter name

the name of the test.

parameter count

number of test cases to run, counting only the test cases which satisfy preconditions.

parameter long_factor

the factor by which to multiply count, max_gen and max_fail when running a long test (default: 1).

parameter max_gen

maximum number of times the generation function is called in total to replace inputs that do not satisfy preconditions (should be >= count).

parameter max_fail

maximum number of failures before we stop generating inputs. This is useful if shrinking takes too much time.

parameter small

kept for compatibility reasons; if provided, replaces the field arbitrary.small. If there is no shrinking function but there is a small function, only the smallest failures will be printed.

parameter if_assumptions_fail

the minimum fraction of tests that must satisfy the precondition for a success to be considered valid. The fraction should be between 0. and 1. A warning will be emitted otherwise if the flag is `Warning, the test will be a failure if the flag is `Fatal. (since 0.10)

val get_law : 'a QCheck2.Test.cell -> 'a -> bool

deprecated

use QCheck2.Test.get_law instead

val get_name : _ QCheck2.Test.cell -> string

deprecated

use QCheck2.Test.get_law instead

deprecated

use QCheck2.Test.get_name instead

val set_name : _ QCheck2.Test.cell -> string -> unit

deprecated

use QCheck2.Test.get_name instead

deprecated

use QCheck2.Test.set_name instead

val get_count : _ QCheck2.Test.cell -> int

Get the count of a cell.

deprecated

use QCheck2.Test.get_count instead

since

0.5.3

val get_long_factor : _ QCheck2.Test.cell -> int

Get the long factor of a cell.

deprecated

use QCheck2.Test.get_long_factor instead

since

0.5.3

val make : ?⁠if_assumptions_fail:([ `Fatal | `Warning ] * float) -> ?⁠count:int -> ?⁠long_factor:int -> ?⁠max_gen:int -> ?⁠max_fail:int -> ?⁠small:('a -> int) -> ?⁠name:string -> 'a arbitrary -> ('a -> bool) -> QCheck2.Test.t

make arb prop builds a test that checks property prop on instances of the generator arb. See make_cell for a description of the parameters.