Module QCheck2.Fn

Utils on generated functions.

since

0.6

val print : 'f fun_ Print.t

print f prints the implementation of generated function f.

The implementation always contains a default case, represented as _.

Note that printing a function before it was called in the test may not print the full implementation.

val apply : 'f fun_ -> 'f

apply f returns the underlying function to be used in tests. This is an alias for deconstructing as documented in fun_.